+++ jobbergate-cli/jobbergate_cli/subapps/applications/app.py
@@ -3,6 +3,7 @@ Provide a ``typer`` app that can interact with Application data in a cruddy mann
 """
 
 import pathlib
+from textwrap import dedent
 from typing import Any, Dict, Optional, cast
 
 import typer
@@ -12,7 +13,12 @@ from jobbergate_cli.exceptions import handle_abort
 from jobbergate_cli.render import StyleMapper, render_list_results, render_single_result, terminal_message
 from jobbergate_cli.requests import make_request
 from jobbergate_cli.schemas import JobbergateContext, ListResponseEnvelope
-from jobbergate_cli.subapps.applications.tools import fetch_application_data, load_default_config, upload_application
+from jobbergate_cli.subapps.applications.tools import (
+    fetch_application_data,
+    load_default_config,
+    save_application_files,
+    upload_application,
+)
 
 
 # TODO: move hidden field logic to the API
@@ -313,4 +319,38 @@ def delete(
         The application was successfully deleted.
         """,
         subject="Application delete succeeded",
+    )
+
+
+@app.command()
+@handle_abort
+def download_files(
+    ctx: typer.Context,
+    id: Optional[int] = typer.Option(
+        None,
+        help=f"The specific id of the application. {ID_NOTE}",
+    ),
+    identifier: Optional[str] = typer.Option(
+        None,
+        help=f"The human-friendly identifier of the application. {IDENTIFIER_NOTE}",
+    ),
+):
+    """
+    Download the files from an application to the current working directory.
+    """
+    jg_ctx: JobbergateContext = ctx.obj
+    result = fetch_application_data(jg_ctx, id=id, identifier=identifier)
+    saved_files = save_application_files(
+        application_data=result,
+        destination_path=pathlib.Path.cwd(),
+    )
+    terminal_message(
+        dedent(
+            """
+            A total of {} application files were successfully downloaded.
+            """.format(
+                len(saved_files)
+            )
+        ).strip(),
+        subject="Application download succeeded",
     )