+++ jobbergate-cli/jobbergate_cli/subapps/job_scripts/tools.py
@@ -6,6 +6,8 @@ import json
 import pathlib
 from typing import Any, Dict, List, Optional, cast
 
+from loguru import logger
+
 from jobbergate_cli.exceptions import Abort
 from jobbergate_cli.requests import make_request
 from jobbergate_cli.schemas import JobbergateContext, JobScriptCreateRequestData, JobScriptResponse
@@ -127,3 +129,23 @@ def create_job_script(
     )
 
     return job_script_result
+
+
+def save_job_script_files(
+    job_script_data: JobScriptResponse,
+    destination_path: pathlib.Path,
+) -> List[pathlib.Path]:
+    """
+    Safe the job script files from the API response to the output path.
+    """
+    logger.debug(f"Saving job script files to {destination_path.as_posix()}")
+    saved_files: List[pathlib.Path] = []
+
+    for filename, file_content in job_script_data.job_script_files.files.items():
+        file_path = destination_path / filename
+        file_path.parent.mkdir(parents=True, exist_ok=True)
+        file_path.write_text(file_content)
+        saved_files.append(file_path)
+
+    logger.debug(f"The following files were saved: {list(map(str, saved_files))}")
+    return saved_files