+++ jobbergate-cli/tests/subapps/job_scripts/test_app.py
@@ -1,8 +1,10 @@
 import importlib
 import json
 import shlex
+from unittest import mock
 
 import httpx
+import pytest
 
 from jobbergate_cli.schemas import (
     ApplicationResponse,
@@ -16,8 +18,10 @@ from jobbergate_cli.subapps.job_scripts.app import (
     JOB_SUBMISSION_HIDDEN_FIELDS,
     create,
     delete,
+    download_files,
     get_one,
     list_all,
+    pathlib,
     show_files,
     style_mapper,
     update,
@@ -399,3 +403,73 @@ def test_show_files__success(
         subject="application.sh",
         footer="This is the main job script file",
     )
+
+
+class TestDownloadJobScriptFiles:
+    """
+    Test the ``download`` subcommand.
+    """
+
+    @pytest.fixture()
+    def test_app(self, make_test_app):
+        """
+        Fixture to create a test app.
+        """
+        return make_test_app("download", download_files)
+
+    def test_download__success(
+        self,
+        respx_mock,
+        test_app,
+        dummy_job_script_data,
+        dummy_domain,
+        cli_runner,
+        mocker,
+        tmp_path,
+        dummy_template_source,
+    ):
+        """
+        Test that the ``download`` subcommand works as expected.
+        """
+        respx_mock.get(f"{dummy_domain}/jobbergate/job-scripts/1").mock(
+            return_value=httpx.Response(
+                httpx.codes.OK,
+                json=dummy_job_script_data[0],
+            ),
+        )
+        mocked_render = mocker.patch("jobbergate_cli.subapps.job_scripts.app.terminal_message")
+
+        with mock.patch.object(pathlib.Path, "cwd", return_value=tmp_path):
+            result = cli_runner.invoke(test_app, shlex.split("download --id=1"))
+
+        assert result.exit_code == 0, f"download failed: {result.stdout}"
+        mocked_render.assert_called_once_with(
+            "A total of 1 job script files were successfully downloaded.",
+            subject="Job script download succeeded",
+        )
+
+        desired_list_of_files = {tmp_path / "application.sh"}
+        assert set(tmp_path.rglob("*")) == set(desired_list_of_files)
+        assert (tmp_path / "application.sh").read_text() == dummy_template_source
+
+    def test_download__fail(
+        self,
+        respx_mock,
+        test_app,
+        dummy_domain,
+        cli_runner,
+    ):
+        """
+        Test that the ``download`` subcommand fails when the job script does not exist.
+        """
+        respx_mock.get(f"{dummy_domain}/jobbergate/job-scripts/1").mock(
+            return_value=httpx.Response(
+                httpx.codes.NOT_FOUND,
+            ),
+        )
+
+        with mock.patch("jobbergate_cli.subapps.job_scripts.tools.save_job_script_files") as mocked:
+            result = cli_runner.invoke(test_app, shlex.split("download --id=1"))
+            mocked.assert_not_called()
+
+        assert result.exit_code == 1